requirements.txt
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0

# Memory system dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM dependencies
openai>=1.0.0
anthropic>=0.5.0

# LangChain for unified LLM interface
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-deepseek>=0.0.1
langchain-community>=0.0.20

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Financial data
yfinance>=0.2.32
pandas-ta>=0.3.14b
finnhub-python>=2.4.0

# Backtesting
vectorbt>=0.28.0
backtrader>=1.9.78.123

# Dashboard
streamlit>=1.50.0
plotly>=6.0.0

# News API
newsapi-python>=0.2.7

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
